# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rarible client, and Service to interact with the Rarible API.",
        "title": "Rarible",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/rarible/ownerships/{id}": {
            "get": {
                "description": "Returns NFT ownership details by ownership ID",
                "tags": [
                    "ownership"
                ],
                "summary": "Get NFT Ownership by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ownership ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RaribleAPI_internal_model.OwnershipResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/rarible/traits/rarity": {
            "post": {
                "description": "Returns rarity for NFT traits (recommended)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rarity"
                ],
                "summary": "Get NFT Traits Rarity",
                "parameters": [
                    {
                        "description": "Rarity request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RaribleAPI_internal_model.RarityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RaribleAPI_internal_model.RarityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "RaribleAPI_internal_model.Asset": {
            "type": "object",
            "properties": {
                "type": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.AssetType"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "RaribleAPI_internal_model.AssetType": {
            "type": "object",
            "properties": {
                "@type": {
                    "type": "string"
                },
                "blockchain": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                }
            }
        },
        "RaribleAPI_internal_model.Order": {
            "type": "object",
            "properties": {
                "cancelled": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.OrderData"
                },
                "dbUpdatedAt": {
                    "type": "string"
                },
                "endedAt": {
                    "type": "string"
                },
                "feeTakers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fill": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lastUpdatedAt": {
                    "type": "string"
                },
                "make": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.Asset"
                },
                "makePrice": {
                    "type": "number"
                },
                "makePriceUsd": {
                    "type": "number"
                },
                "makeStock": {
                    "type": "number"
                },
                "maker": {
                    "type": "string"
                },
                "optionalRoyalties": {
                    "type": "boolean"
                },
                "platform": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "take": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.Asset"
                },
                "takePrice": {
                    "type": "number"
                },
                "takePriceUsd": {
                    "type": "number"
                },
                "taker": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "RaribleAPI_internal_model.OrderData": {
            "type": "object",
            "properties": {
                "@type": {
                    "type": "string"
                },
                "data": {}
            }
        },
        "RaribleAPI_internal_model.OriginOrder": {
            "type": "object",
            "properties": {
                "bestBidOrder": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.Order"
                },
                "bestSellOrder": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.Order"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "RaribleAPI_internal_model.OwnershipResponse": {
            "type": "object",
            "properties": {
                "bestSellOrder": {
                    "$ref": "#/definitions/RaribleAPI_internal_model.Order"
                },
                "blockchain": {
                    "type": "string"
                },
                "collection": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itemId": {
                    "type": "string"
                },
                "lastUpdatedAt": {
                    "type": "string"
                },
                "lazyValue": {
                    "type": "integer"
                },
                "originOrders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RaribleAPI_internal_model.OriginOrder"
                    }
                },
                "owner": {
                    "type": "string"
                },
                "pending": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RaribleAPI_internal_model.Pending"
                    }
                },
                "source": {
                    "type": "string"
                },
                "tokenId": {
                    "type": "integer"
                },
                "value": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "RaribleAPI_internal_model.Pending": {
            "type": "object",
            "properties": {
                "@type": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "royalties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RaribleAPI_internal_model.Royalty"
                    }
                }
            }
        },
        "RaribleAPI_internal_model.RarityRequest": {
            "type": "object",
            "properties": {
                "collection": {
                    "type": "string"
                }
            }
        },
        "RaribleAPI_internal_model.RarityResponse": {
            "type": "object",
            "properties": {
                "continuation": {
                    "type": "string"
                },
                "traits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RaribleAPI_internal_model.RarityTrait"
                    }
                }
            }
        },
        "RaribleAPI_internal_model.RarityTrait": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "rarity": {
                    "type": "number"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "RaribleAPI_internal_model.Royalty": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        }
    }
}